crawler/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pytest>=7.4.0
pytest-mock>=3.12.0
opencc-python-reimplemented>=0.1.7
